server/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
apscheduler==3.11.0
byted-project-oapi-sdk==0.0.5
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
colorlog==6.9.0
fastapi==0.116.1
fastapi-mqtt==2.2.0
gmqtt==0.7.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
lark-oapi==1.4.19
packaging==25.0
pip==25.1.1
pluggy==1.6.0
prometheus-client==0.22.1
prometheus-fastapi-instrumentator==7.1.0
psutil==7.0.0
pycryptodome==3.23.0
pydantic==2.11.7
pydantic-core==2.33.2
pygments==2.19.2
pytest==8.4.1
pytest-asyncio==1.0.0
python-dotenv==1.1.1
pyyaml==6.0.2
requests==2.31.0
requests-toolbelt==1.0.0
ruamel-yaml==0.18.14
ruamel-yaml-clib==0.2.12
sniffio==1.3.1
sqlalchemy==2.0.41
starlette==0.47.1
typing-extensions==4.14.0
typing-inspection==0.4.1
tzlocal==5.3.1
urllib3==2.5.0
uvicorn==0.35.0
websockets==15.0.1
